requirements.txt
flask>=3.0.0
openai>=1.40.0
httpx>=0.27.0
python-dotenv>=1.0.0
numpy>=1.24.0
faiss-cpu>=1.8.0
sentence-transformers>=2.7.0
torch>=2.0.0
tqdm>=4.66.0
faster-whisper>=1.0.0
